Rename: request와 response 분리
