+++ src/main/java/io/ejangs/docsa/domain/doc/dto/response/DocCreateResponse.java
@@ -1,4 +1,4 @@
-package io.ejangs.docsa.domain.doc.dto;
+package io.ejangs.docsa.domain.doc.dto.response;
 
 public record DocCreateResponse(
         Long id
